+++ libyul/Object.h
@@ -26,6 +26,7 @@
 #include <libdevcore/Common.h>
 
 #include <memory>
+#include <set>
 
 namespace yul
 {
@@ -62,6 +63,10 @@ struct Object: ObjectNode
 public:
 	/// @returns a (parseable) string representation. Includes types if @a _yul is set.
 	std::string toString(bool _yul) const override;
+
+	/// @returns the set of names of data objects accessible from within the code of
+	/// this object.
+	std::set<YulString> dataNames() const;
 
 	std::shared_ptr<Block> code;
 	std::vector<std::shared_ptr<ObjectNode>> subObjects;