+++ libyul/optimiser/StackCompressor.cpp
@@ -155,19 +155,20 @@ void eliminateVariables(
 
 bool StackCompressor::run(
 	Dialect const& _dialect,
-	Block& _ast,
+	Object& _object,
 	bool _optimizeStackAllocation,
 	size_t _maxIterations
 )
 {
 	yulAssert(
-		_ast.statements.size() > 0 && _ast.statements.at(0).type() == typeid(Block),
+		_object.code &&
+		_object.code->statements.size() > 0 && _object.code->statements.at(0).type() == typeid(Block),
 		"Need to run the function grouper before the stack compressor."
 	);
-	bool allowMSizeOptimzation = !SideEffectsCollector(_dialect, _ast).containsMSize();
+	bool allowMSizeOptimzation = !SideEffectsCollector(_dialect, *_object.code).containsMSize();
 	for (size_t iterations = 0; iterations < _maxIterations; iterations++)
 	{
-		map<YulString, int> stackSurplus = CompilabilityChecker::run(_dialect, _ast, _optimizeStackAllocation);
+		map<YulString, int> stackSurplus = CompilabilityChecker::run(_dialect, _object, _optimizeStackAllocation);
 		if (stackSurplus.empty())
 			return true;
 
@@ -176,15 +177,15 @@ bool StackCompressor::run(
 			yulAssert(stackSurplus.at({}) > 0, "Invalid surplus value.");
 			eliminateVariables(
 				_dialect,
-				boost::get<Block>(_ast.statements.at(0)),
+				boost::get<Block>(_object.code->statements.at(0)),
 				stackSurplus.at({}),
 				allowMSizeOptimzation
 			);
 		}
 
-		for (size_t i = 1; i < _ast.statements.size(); ++i)
+		for (size_t i = 1; i < _object.code->statements.size(); ++i)
 		{
-			FunctionDefinition& fun = boost::get<FunctionDefinition>(_ast.statements[i]);
+			FunctionDefinition& fun = boost::get<FunctionDefinition>(_object.code->statements[i]);
 			if (!stackSurplus.count(fun.name))
 				continue;